refactored styles, added favicon
