+++ src/views/GoHome/GoHomeStyle.js
